feat: now use single thread for many stuffs

+++ src/hcf/object/faction/query/DisbandFactionQuery.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace hcf\object\faction\query;
 
-use hcf\thread\datasource\MySQL;
-use hcf\thread\datasource\Query;
+use hcf\thread\LocalThreaded;
+use hcf\thread\types\SQLDataSourceThread;
+use hcf\thread\types\ThreadType;
 
-final class DisbandFactionQuery extends Query {
+final class DisbandFactionQuery implements LocalThreaded {
 
     /**
      * @param string $id
@@ -15,11 +16,22 @@ final class DisbandFactionQuery extends Query {
     public function __construct(private string $id) {}
 
     /**
-     * @param MySQL $provider
+     * @return int
+     */
+    public function threadId(): int {
+        return 0;
+    }
+
+    /**
+     * @param ThreadType $threadType
      *
      * This function is executed on other Thread to prevent lag spike on Main thread
      */
-    public function run(MySQL $provider): void {
+    public function run(ThreadType $threadType): void {
+        if (!$threadType instanceof SQLDataSourceThread || $threadType->id() !== $this->threadId()) return;
+
+        $provider = $threadType->getResource();
+
         $provider->prepareStatement('DELETE FROM factions WHERE id = ?');
         $provider->set($this->id);
